write params and buffers in param_string

+++ src/results/mod.rs
@@ -1,17 +1,16 @@
-
 pub mod xml_reader;
 // pub mod csv_writer;
 
 #[derive(Debug, Serialize, Deserialize, PartialEq)]
 pub struct ResultRow {
-    pass: u32,
-    result: f32,
+    pass: u64,
+    result: f64,
     profit: f32,
     expected_payoff: f32,
     profit_factor: f32,
     recovery_factor: f32,
     sharpe_ratio: f32,
-    custom: f32,
+    custom: f64,
     equity_dd: f32,
     trades: u32,
     params: Vec<f32>,
